Copter: create autoyaw.cpp from all the bits of Copter::Mode::AutoYaw
